new gl=angle-egl option
