+++ packages/docs/docs/cli/still.md
@@ -107,20 +107,7 @@ Opens an actual browser during rendering to observe the render.
 
 ### `--gl`
 
-<AngleChangelog />
-
-Select the OpenGL renderer backend for Chromium.
-Accepted values:
-
-- `"angle"`,
-- `"egl"`,
-- `"swiftshader"`
-- `"swangle"`
-- `"vulkan"` (_from Remotion v4.0.41_)
-- `null` - Chromium's default
-
-**Default for local rendering**: `null`.  
-**Default for Lambda rendering**: `"swangle"`.
+<Options id="gl" cli />
 
 ### `--user-agent`<AvailableFrom v="3.3.83"/>
