+++ packages/docs/docs/renderer/render-frames.md
@@ -275,20 +275,7 @@ If disabled, the render will open an actual Chrome window where you can see the 
 
 _string_
 
-<AngleChangelog />
-
-Select the OpenGL renderer backend for Chromium.
-Accepted values:
-
-- `"angle"`,
-- `"egl"`,
-- `"swiftshader"`
-- `"swangle"`
-- `"vulkan"` (_from Remotion v4.0.41_)
-- `null` - Chromiums default
-
-**Default for local rendering**: `null`.  
-**Default for Lambda rendering**: `"swangle"`.
+<Options id="gl" />
 
 #### `userAgent`<AvailableFrom v="3.3.83"/>
